feat(pagination): fix to async logic
